requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
streamlit==1.45.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.20
langchain-core==0.3.59
langchain-community==0.3.19
langchain-google-genai==2.1.2
langgraph==0.3.21
langgraph-prebuilt==0.1.7
google-genai==1.8.0
google-api-core==2.24.1
google-auth==2.38.0
typing-extensions==4.12.2
python-dateutil==2.9.0.post0
requests==2.32.3
numpy==1.26.4
pandas==2.2.1 
sqlite3==3.27.2
